remove prompts from pull secrets
